requirements.txt
transformers
textblob
nltk
scikit-learn
vaderSentiment
textstat
matplotlib
seaborn
python-docx
fpdf
